requirements.txt
numpy>=1.21.0
matplotlib>=3.4.0
pandas>=1.3.0
seaborn>=0.11.0
tsplib95>=0.7.1
tabulate>=0.8.9
tqdm>=4.62.0
